chore: merge wallet_details
