+++ src/renderer/widgets/WalletDetails/model/wallet-provider-model.ts
@@ -1,4 +1,4 @@
-import { combine, createEvent, createStore, forward, sample } from 'effector';
+import { combine } from 'effector';
 
 import { accountUtils, walletModel } from '@renderer/entities/wallet';
 import { walletSelectModel } from '@renderer/features/wallets';
@@ -14,14 +14,6 @@ import type {
   ChainAccount,
   ChainId,
 } from '@renderer/shared/core';
-import { Account, Wallet } from '@renderer/shared/core';
-import { ForgetStep } from '../lib/constants';
-
-const reset = createEvent();
-const forgetButtonClicked = createEvent();
-const forgetModalClosed = createEvent();
-
-const $forgetStep = createStore<ForgetStep>(ForgetStep.NOT_STARTED).reset(reset);
 
 const $accounts = combine(
   {
@@ -126,39 +118,8 @@ const $isConnected = combine(
   },
 );
 
-sample({
-  clock: forgetButtonClicked,
-  fn: () => ForgetStep.FORGETTING,
-  target: $forgetStep,
-});
-
-sample({
-  clock: forgetButtonClicked,
-  source: walletSelectModel.$walletForDetails,
-  filter: (wallet): wallet is Wallet => wallet !== null,
-  fn: (wallet) => wallet!.id,
-  target: walletModel.events.walletRemoved,
-});
-
-sample({
-  clock: walletModel.events.walletRemovedSuccess,
-  fn: () => ForgetStep.SUCCESS,
-  target: $forgetStep,
-});
-
-forward({
-  from: forgetModalClosed,
-  to: walletSelectModel.events.walletForDetailsCleared,
-});
-
 export const walletProviderModel = {
   $accounts,
-  $forgetStep,
-  events: {
-    reset,
-    forgetButtonClicked,
-    forgetModalClosed,
-  },
   $singleShardAccount,
   $multiShardAccounts,
   $multisigAccount,
